requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-filter==2.4.0
markdown==3.3.3
ipython==7.16.1
pytest==6.2.1
pytest-django==4.1.0
autopep8==1.5.4
